filehunter/requirements.txt
customtkinter
requests
pillow
pymupdf